pkgdiff-mg: under systemd use /var/tmp as default TMPDIR
pkgdiff can cause large files in the used temprorary directory that
are not cleaned up on exit. Using /tmp for those, which is often a
tmpfs, which means that the files consume main memory at first. Later
the system may move the unused files on the tmpfs into swap.

Under systemd, using /var/tmp for files that are not deleted by the
creator, which is pkgdiff-mg in our case, is safe, since systemd will
automatically delete old unused files from /var/tmp.

Signed-off-by: Florian Schmaus <flow@gentoo.org>